Move some EditAlgTest tests and all SelectorAlgTest to RewriteTest
